setting/#1: 경로 수정 및 파일 제거
