+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:usesCleartextTraffic="true"
         android:theme="@style/Theme.Kiero">
         <activity
-            android:name="com.kiero.presentation.main.MainActivity"
+            android:name="com.kiero.presentation.main.activity.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Kiero">
