--- app/src/main/java/com/kiero/core/navigation/MainTabRoute.kt
@@ -1,3 +1,0 @@
-package com.kiero.core.navigation
-
-interface MainTabRoute : Route
