+++ app/src/main/java/com/kiero/presentation/main/activity/MainActivity.kt
@@ -1,0 +1,23 @@
+package com.kiero.presentation.main.activity
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import com.kiero.core.designsystem.theme.KieroTheme
+import com.kiero.presentation.main.screen.MainRoute
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+
+        setContent {
+            KieroTheme {
+                MainRoute()
+            }
+        }
+    }
+}
